Modularized getting redirect URL. Moved getSongs to proper location.
